Move `pendingBridgeTransactions` actions to `transactions`
